backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-socketio==5.11.1
langgraph==0.2.34
langchain==0.3.7
langchain-openai>=0.2.1
playwright==1.48.0
pydantic==2.9.2
python-dotenv==1.0.1
aiohttp==3.10.10
jinja2==3.1.4
httpx==0.27.2
sqlalchemy==2.0.35
asyncpg==0.29.0
redis==5.0.3
langgraph-checkpoint-redis==0.1.0
